Adjust page title; Design sidebar for mobile devices
Change the design of the title box (around the name and job title) to
three lines rather than a box with thick lines on the sides and a thin
line on top and bottom.  This was purely asthetic, and I may go back and
provide the means to allow the end user determine what style they prefer
at some future date.

Add 'Contact' section to Sidebar containing phone and email.  Mailing
address was omitted because users maay not want this information
displayed readily on the internet.

Add support for small low resolution screens and mobile devices with the
sidebar.  This section will now shift to a horizontal layout across
the page between the header and the 'Experience' section with the
'Professional Skills', 'Personal Talents', and 'Spoken Languages' fields
collapsed on any screen smaller than 1,300px.  A plus/minus sign
on the right hand side is used to expand/collapse these sections.