Add layout for 'Experience'
Add 'Experience' section to the resume, where users can list past
jobs.  Layout is:

   COMPANY                                              City, ST
   title                                                   dates
      job description.....................................
      ....................................................
      ....................................................

Modify skillTalent.js to require parameter for parent class and child
class.  parentClass contains the class name for the first parent
element in the DOM structure that contains both the button used to
initiate hiding a div and the div to be hiden.  childClass contains the
class name for the div to be hidden.  Technically, an ID or any valid
JQuery selector could be passed into this routine.